refactoring - added grade component, user model and user and grade services
