+++ frontend/src/App.tsx
@@ -1,54 +1,19 @@
 import React from 'react';
-import Counter from './components/counter/Counter';
+import ManageUser from './components/ManageUser/ManageUser';
 import './App.css';
+import {
+    BrowserRouter as Router,
+    Route,
+} from "react-router-dom";
+import ManageGrades from "./components/ManageGrades";
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <Counter />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <span>
-          <span>Learn </span>
-          <a
-            className="App-link"
-            href="https://reactjs.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React
-          </a>
-          <span>, </span>
-          <a
-            className="App-link"
-            href="https://redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux
-          </a>
-          <span>, </span>
-          <a
-            className="App-link"
-            href="https://redux-toolkit.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Redux Toolkit
-          </a>
-          ,<span> and </span>
-          <a
-            className="App-link"
-            href="https://react-redux.js.org/"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            React Redux
-          </a>
-        </span>
-      </header>
+        <Router>
+            <Route path="/:userID" component={ManageUser} />
+            <Route path="/grades" component={ManageGrades} />
+        </Router>
     </div>
   );
 }
