+++ frontend/src/components/ManageUser/ManageUser.module.css
@@ -37,7 +37,6 @@ body {
   font-size: 1.1rem;
   text-align: left;
   margin: 1.5rem 5rem 2rem 2rem;
-  border: magenta 1px solid;
 }
 .manageUserForm__row {
   display: flex;
@@ -65,7 +64,6 @@ body {
   border: #fff 1px solid;
 }
 .manageUserForm__row__value__grades {
-  border: magenta 1px solid;
   text-align: right;
 }
 .manageUserForm__row__value__grades input {
