+++ frontend/src/components/ManageUser/ManageUser.module.scss
@@ -47,7 +47,6 @@ body {
   font-size: 1.1rem;
   text-align: left;
   margin: 1.5rem 5rem 2rem 2rem;
-  border: magenta 1px solid;
 
   &__row {
     display: flex;
@@ -76,7 +75,6 @@ body {
         }
       }
       &__grades {
-        border: magenta 1px solid;
         text-align: right;
         input {
           padding: .7rem 1rem;
